llm-hackathon-ingestion/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
boto3==1.28.28
    # via -r requirements.in
botocore==1.31.28
    # via
    #   boto3
    #   s3transfer
certifi==2023.7.22
    # via requests
charset-normalizer==3.0.1
    # via requests
html2text==2024.2.26
    # via -r requirements.in
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
requests==2.31.0
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.14
    # via
    #   botocore
    #   requests
